Wrapper to provide .NET ILogger style functions

+++ graylog.go
@@ -37,7 +37,7 @@ type Message struct {
 	Host         string            `json:"host"`
 	ShortMessage string            `json:"short_message"`
 	FullMessage  string            `json:"full_message,omitempty"`
-	Timestamp    int64             `json:"timestamp,omitempty"`
+	Timestamp    float64           `json:"timestamp,omitempty"`
 	Level        uint              `json:"level,omitempty"`
 	Extra        map[string]string `json:"-"`
 }
